refactor structure of the repo
